Version 1.3
Add PyInstaller, quick debug packaging, shortcuts for both py2 and py3 installation at the same time. Shorten long lines. Visual update.

+++ kivy.bat
@@ -1,5 +1,5 @@
 ::Author: KeyWeeUsr @ https://github.com/KeyWeeUsr
-::Version: 1.2.1
+::Version: 1.3
 ::Inspired by kivy.bat file for kivy1.8.0
 ::To reset file just delete "config.kivyinstaller"
 ::Bitsadmin is available since winXP SP2
@@ -20,9 +20,10 @@ set pyversion=0
 set gstreamer=0
 set master=1.9.2
 set installkivy=1
-set installerversion=1.2.1
+set installerversion=1.3
 setlocal ENABLEDELAYEDEXPANSION
 ver | find "5.1" >nul && set xp=1
+title = KivyInstaller %installerversion%
 set sendto=%appdata%\Microsoft\Windows\SendTo
 set taskbar=%appdata%\Microsoft\Internet Explorer\Quick Launch
 set addlocal=ADDLOCAL^=DefaultFeature,PrivateCRT,TclTk,Documentation,Tools,Testsuite
@@ -30,12 +31,12 @@ if %xp%==1 (
     set sendto=%userprofile%\SendTo
 )
 
-echo #####################################################################
+echo ###############################################################################
 echo KivyInstaller v%installerversion%
 echo Author: KeyWeeUsr @ https://github.com/KeyWeeUsr
 echo.
 echo Report issues @ https://github.com/KeyWeeUsr/KivyInstaller/issues
-echo #####################################################################
+echo ###############################################################################
 
 echo Looking for config file...
 if exist "%~dp0config.kivyinstaller" (
@@ -72,10 +73,12 @@ if %choice%==2 (
     set pyversion=%py2%
     set cp=%cp2%m
     set cpwhl=%cp2%
+    set shrtct=%cp2:~2%
 ) else if %choice%==3 (
     set pyversion=%py3%
     set cp=%cp3%m
     set cpwhl=%cp3%
+    set shrtct=%cp3:~2%
 ) else (
     goto version
 )
@@ -139,39 +142,45 @@ goto check
 
 :uninstall
 set /p choice="Uninstall? y/n"
+set /p choice2="Remove cached pip files? y/n"
+set /p choice3="Remove dist folder? y/n"
 if %arch%==win32 (
     if not exist "%~dp0py%pyversion%.msi" (
         bitsadmin.exe /transfer "GetPythonMSI" "https://www.python.org/ftp/python/%pyversion%/python-%pyversion%.msi" "%~dp0py%pyversion%.msi"
     )
     if %choice%==y (
         msiexec.exe /x "%~dp0py%pyversion%.msi" /qb
-    )
+    ) else (goto end)
 ) else (
     if not exist "%~dp0py%pyversion%.amd64.msi" (
         bitsadmin.exe /transfer "GetPythonMSI" "https://www.python.org/ftp/python/%pyversion%/python-%pyversion%.amd64.msi" "%~dp0py%pyversion%.amd64.msi"
     )
     if %choice%==y (
         msiexec.exe /x "%~dp0py%pyversion%.amd64.msi" /qb
-    )
+    ) else (goto end)
 )
 if not exist "%~dp0python.exe" (
     del /q "%~dp0config.kivyinstaller"
     del /q "%~dp0update_kivy.bat"
     del /q "%~dp0backup_kivy.bat"
+    del /s /q /f "%~dp0*.spec"
     rmdir /s /q "%userprofile%\.kivy"
     rmdir /s /q "%userprofile%\.idlerc"
     rmdir /s /q "%~dp0Lib"
     rmdir /s /q "%~dp0Scripts"
     rmdir /s /q "%~dp0share"
     rmdir /s /q "%~dp0whls"
+    rmdir /s /q "%~dp0build"
 )
-set /p choice="Remove cached pip files? y/n"
-if %choice%==y (
+if %choice2%==y (
     if exist "%localappdata%\pip" (
         rmdir /s /q "%localappdata%\pip"
     ) else (
         rmdir /s /q "%userprofile%\Local Settings\Application Data\pip"
     )
+)
+if %choice3%==y (
+    rmdir /s /q "%~dp0dist"
 )
 goto rmshortcuts
 
@@ -202,9 +211,9 @@ if %installerversion% lss %updateversion% (
     goto console
 )
 echo Backing up original file...
-echo f | xcopy "%~dp0%~n0.bat" "%~dp0backup_kivy.bat"
+echo F | xcopy /y "%~dp0%~n0.bat" "%~dp0backup_kivy.bat"
 echo Updating...
-start cmd /c timeout /t 5 & (echo Y | xcopy "%~dp0update_kivy.bat" "%~dp0%~n0.bat") & del %~dp0update_kivy.bat && cmd /c "%~dp0%~n0"
+start cmd /c timeout /t 5 & (echo Y | xcopy "%~dp0update_kivy.bat" "%~dp0%~n0.bat") & del "%~dp0update_kivy.bat"
 exit
 
 :help
@@ -214,19 +223,21 @@ echo.
 echo Usage:
 echo   %~n0 [options]
 echo.
-echo   ^<file^>             Run python(.py) file.
-echo   update             Update kivy wheel to the latest stable.
-echo   updatemaster       Update kivy wheel to the latest nightly-build.
-echo   batcheck           Check for new KivyInstaller version only.
-echo   batupdate          Check for new KivyInstaller version and update.
-echo   remove             Uninstall kivy only.
-echo   uninstall          Uninstall kivy, python and leave only %~n0.bat.
-echo   mkshortcuts        Create shortcuts for SendTo and TaskBar.
-echo   rmshortcuts        Remove shortcuts for SendTo and TaskBar.
-echo   help               Show this.
+echo   ^<file^>                  Run python(.py) file.
+echo   update                  Update kivy wheel to the latest stable.
+echo   updatemaster            Update kivy wheel to the latest nightly-build.
+echo   batcheck                Check for new KivyInstaller version only.
+echo   batupdate               Check for new KivyInstaller version and update.
+echo   remove                  Uninstall kivy only.
+echo   uninstall               Uninstall kivy, python and leave only %~n0.bat.
+echo   mkshortcuts             Create shortcuts for SendTo and TaskBar.
+echo   rmshortcuts             Remove shortcuts for SendTo and TaskBar.
+echo   pack "<path to .py>"    Quick packaging with pyinstaller.
+echo   help                    Show this.
 echo.
 echo Optional:
-echo   idle               Python(.py) file editor with syntax highlighting
+echo   idle                    Python(.py) text editor with syntax highlighting
+echo   pyinstaller             Executable packager for Python programs
 goto end
 
 :check
@@ -240,10 +251,12 @@ if %installkivy%==0 (
 )
 echo Preparing Python for Kivy...
 python -m pip install --upgrade pip wheel setuptools
+set packurl=--extra-index-url https://kivy.org/downloads/packages/simple/
+set packages=docutils pygments pypiwin32 requests wget kivy.deps.sdl2 kivy.deps.glew pyinstaller
 if %gstreamer%==1 (
-    python -m pip install docutils pygments pypiwin32 requests wget kivy.deps.sdl2 kivy.deps.glew kivy.deps.gstreamer --extra-index-url https://kivy.org/downloads/packages/simple/
+    python -m pip install %packages% kivy.deps.gstreamer %packurl%
 ) else (
-    python -m pip install docutils pygments pypiwin32 requests wget kivy.deps.sdl2 kivy.deps.glew --extra-index-url https://kivy.org/downloads/packages/simple/
+    python -m pip install %packages% %packurl%
 )
 if %stable%==1 (
     python -m pip uninstall -y kivy
@@ -251,15 +264,16 @@ if %stable%==1 (
     goto kivyend
 )
 mkdir "%~dp0whls"
+set driveurl=https://drive.google.com/folderview?id=0B1_HB9J8mZepOV81UHpDbmg5SWM^^^&usp=sharing#list
 echo.import os,glob,sys,re,requests,wget,datetime,shutil> "%~dp0getnightly.py"
-echo.nightly = requests.get('https://drive.google.com/folderview?id=0B1_HB9J8mZepOV81UHpDbmg5SWM^&usp=sharing#list').content>> "%~dp0getnightly.py"
->>"%~dp0getnightly.py" echo.yesterday = '{:%%d%%m%%Y}'.format(datetime.datetime.now()-datetime.timedelta(days=1))
+echo.n = requests.get('%driveurl%').content>> "%~dp0getnightly.py"
+>>"%~dp0getnightly.py" echo.y = '{:%%d%%m%%Y}'.format(datetime.datetime.now()-datetime.timedelta(days=1))
 echo.try:>> "%~dp0getnightly.py"
->>"%~dp0getnightly.py" echo.    match = re.findall('''\"(Kivy-\d\.\d\.\d)(\.\w{4}_'''+yesterday+'''_git\_?\w{7}-%cpwhl%)(-none|_\d{8}_git_\w{7}-%cp%)(-%arch%.whl)",,,,,"(\S+)",,\"''',nightly.decode('utf-8'))[0]
+>>"%~dp0getnightly.py" echo.    m=re.findall('''\"(Kivy-\d\.\d\.\d)(\.\w{4}_'''+y+'''_git\_?\w{7}-%cpwhl%)(-none|_\d{8}_git_\w{7}-%cp%)(-%arch%.whl)",,,,,"(\S+)",,\"''',n.decode('utf-8'))[0]
 echo.except IndexError:>> "%~dp0getnightly.py"
 echo.    print("No nightly-build yet")>> "%~dp0getnightly.py"
-echo.file_name = ''.join(match[:len(match)-1])>> "%~dp0getnightly.py"
-echo.file_id = match[-1]>> "%~dp0getnightly.py"
+echo.file_name = ''.join(m[:len(m)-1])>> "%~dp0getnightly.py"
+echo.file_id = m[-1]>> "%~dp0getnightly.py"
 echo.link = 'https://docs.google.com/uc?id='+file_id>> "%~dp0getnightly.py"
 echo.p = os.path.dirname(os.path.abspath(__file__))+'\\whls\\'+file_name.replace('%cp%','none')>> "%~dp0getnightly.py"
 echo.try:>>"%~dp0getnightly.py"
@@ -306,24 +320,40 @@ if not %errorlevel%==1 (
     (echo gstreamer=%gstreamer%) >> "%~dp0config.kivyinstaller"
     (echo master=%master%) >> "%~dp0config.kivyinstaller"
     (echo installkivy=%installkivy%) >> "%~dp0config.kivyinstaller"
+    (echo shrtct=%shrtct%) >> "%~dp0config.kivyinstaller"
 ) else (
     echo Kivy was not installed properly!
 )
 
 :mkshortcuts
 if %shortcuts%==1 (
-    (echo @echo off) > "%taskbar%\Kivy.bat"
-    (echo if "%%~1"=="" ^() >> "%taskbar%\Kivy.bat"
-    (echo     start "" call "%~dp0%~n0.bat") >> "%taskbar%\Kivy.bat"
-    (echo ^) else ^() >> "%taskbar%\Kivy.bat"
-    (echo     start "" call "%~dp0%~n0.bat" "%%~1") >> "%taskbar%\Kivy.bat"
-    (echo ^)) >> "%taskbar%\Kivy.bat" && type "%taskbar%\Kivy.bat" > "%sendto%\Kivy.bat"
+    (echo @echo off) > "%taskbar%\Kivy%shrtct%.bat"
+    (echo if "%%~1"=="" ^() >> "%taskbar%\Kivy%shrtct%.bat"
+    (echo     start "" call "%~dp0%~n0.bat") >> "%taskbar%\Kivy%shrtct%.bat"
+    (echo ^) else ^() >> "%taskbar%\Kivy%shrtct%.bat"
+    (echo     start "" call "%~dp0%~n0.bat" "%%~1") >> "%taskbar%\Kivy%shrtct%.bat"
+    (echo ^)) >> "%taskbar%\Kivy%shrtct%.bat" && type "%taskbar%\Kivy%shrtct%.bat" > "%sendto%\Kivy%shrtct%.bat"
 )
 goto end
 
 :rmshortcuts
-del /q "%sendto%\Kivy.bat"
-del /q "%taskbar%\Kivy.bat"
+del /q "%sendto%\Kivy%shrtct%.bat"
+del /q "%taskbar%\Kivy%shrtct%.bat"
+goto end
+
+:pack
+for /f %%a in ("%2") do for %%b in ("%%~dpa\.") do set n=%%~nxb
+for /f %%a in ("%2") do for %%b in ("%%~dpa") do set d=%%b
+echo Collecting data...
+python -m PyInstaller --debug --name "%n%" "%2"
+echo Editing .spec file...
+set d=%d:\=\\\\%
+set f=from kivy.deps import sdl2, glew\na = 
+set t=a.datas,*[Tree(p) for p in (sdl2.dep_bins + glew.dep_bins)],
+python -c "o=open;f=o('%n%.spec');t=f.read();f.close();f=o('%n%.spec','w');f.write(t.replace('\na = ','%f%').replace('a.datas,','%t%').replace('T(exe,','T(exe,Tree(\'%d:"=%\'),'));f.close();"
+echo Packaging...
+python -m PyInstaller "%n%.spec"
+del /q "%~dp0%n%.spec"
 goto end
 
 :installed
@@ -348,11 +378,13 @@ if [%1]==[update] (
     goto mkshortcuts
 ) else if [%1]==[rmshortcuts] (
     goto rmshortcuts
+) else if [%1]==[pack] (
+    goto pack
 ) else if [%1]==[help] (
     goto help
 )
 cls
-set PATH=%~dp0;%~dp0Tools;%~dp0Lib\idlelib;%PATH%
+set PATH=%~dp0;%~dp0Tools;%~dp0Scripts;%~dp0Lib\idlelib;%PATH%
 echo PATH:
 echo %PATH%
 if [%1]==[] goto console
@@ -360,11 +392,14 @@ echo.
 echo Running "python.exe %*"
 python %*
 if not %errorlevel%==0 (pause)
-goto end
+
+:end
+pause
+exit
 
 :console
 echo.
-echo #####################################################################
+echo ###############################################################################
 @echo OFF
 python -c "import sys; print('.'.join(str(x) for x in sys.version_info[:3]))" > temp.txt
 set /p python_version=<temp.txt
@@ -372,21 +407,18 @@ del temp.txt
 start "" /min /wait python.exe -c "from kivy import __version__;f=open('temp.txt','w');f.write(str(__version__)[:5]);f.close()"
 set /p kivy_version=<temp.txt
 del temp.txt
-python -c "exec(\"try:\n    import os, glob, re, datetime, sys;p=os.path.dirname(os.path.abspath(sys.executable));match = re.findall('(\d{8})', os.path.basename(max(glob.glob(p+'/whls/*.[Ww][Hh][Ll]*'), key=os.path.getctime))[:-4])[0];print(match[:2]+' '+datetime.date(1900, int(match[2:4]), 1).strftime('%%B')[:3]+' '+match[4:]);\nexcept:\n    pass\")" >temp.txt
-set /p wheel_version=<temp.txt
-del temp.txt
+python -c "exec(\"try:\n    import os.path as o,glob,re,datetime,sys;    p=o.dirname(o.abspath(sys.executable));    m=re.findall('(\d{8})', o.basename(max(glob.glob(p+'/whls/*.[Ww][Hh][Ll]*'), key=o.getctime))[:-4])[0];    print(m[:2]+datetime.date(1900, int(m[2:4]), 1).strftime('%%B')[:3]+m[4:]);\nexcept:\n    pass\")">t
+set /p wheel_version=<t
+del t
 echo - KivyInstaller: %installerversion%
-echo - Python: %python_version%
-echo - Kivy: %kivy_version%
-echo - Wheel: %wheel_version%
-echo - New wheel: %~n0 update or %~n0 updatemaster
-echo - Examples: share\kivy-examples
-echo - Launch: %~n0 main.py or python main.py
-echo - Uninstall: %~n0 uninstall
-echo #####################################################################
+echo - Python:        %python_version%
+echo - Kivy:          %kivy_version%
+echo - Wheel:         %wheel_version%
+echo - New wheel:     %~n0 update or %~n0 updatemaster
+echo - Examples:      share\kivy-examples
+echo - Launch:        %~n0 main.py or python main.py
+echo - Pack:          %~n0 pack "<path>"
+echo - Uninstall:     %~n0 uninstall
+echo ###############################################################################
 echo.
 cmd
-
-:end
-pause
-exit
